AudioLengthPath: move unit test in parallel to code
